# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted, multi-tenant authentication and authorization service. Provides JWT-based auth, RBAC, API keys, audit logging, and webhook event delivery.",
        "title": "UniAuth API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "UniAuth Support",
            "url": "https://github.com/osama1998h/uniauth"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/email/verify-confirm": {
            "post": {
                "description": "Validates the one-time verification token and marks the user's email as verified.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Confirm email verification",
                "parameters": [
                    {
                        "description": "Verification token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.VerifyEmailConfirmBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token invalid or expired",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/email/verify-request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends a verification email to the authenticated user. Returns 409 if the email is already verified.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Request email verification",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerMessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Account is inactive",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email already verified",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "description": "Validates credentials and returns a JWT access token and a refresh token. Only the access token is accepted in the Authorization header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.TokenPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "User or organization inactive",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Rate limiting temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes the provided refresh token and blacklists the current access token. Authorization must use an access token; the request body must contain the refresh token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout current session",
                "parameters": [
                    {
                        "description": "Refresh token to revoke",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Authentication temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout-all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes all active sessions for the authenticated user and blacklists the current access token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout all sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerMessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Authentication temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/password/change": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Changes the authenticated user's password. Requires the current password for verification.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Change password",
                "parameters": [
                    {
                        "description": "Current and new passwords",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ChangePasswordBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Wrong current password",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Authentication temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/password/reset-confirm": {
            "post": {
                "description": "Validates the one-time reset token and sets a new password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Confirm a password reset",
                "parameters": [
                    {
                        "description": "Reset token and new password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ResetConfirmBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token invalid or expired",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Rate limiting temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/password/reset-request": {
            "post": {
                "description": "Attempts password reset email delivery via configured SMTP. Always returns 200 to prevent email enumeration, even when delivery is unavailable.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Request a password reset email",
                "parameters": [
                    {
                        "description": "Organization slug and email",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ResetRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Rate limiting temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api_handlers.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@acme.com"
                },
                "org_slug": {
                    "type": "string",
                    "example": "acme-corp"
                },
                "password": {
                    "type": "string",
                    "example": "S3cur3P@ss!"
                }
            }
        },
        "internal_api_handlers.LogoutRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGci..."
                }
            }
        },
...
        "internal_api_handlers.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
        "internal_api_handlers.SwaggerMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation completed successfully"
                }
            }
        },
        "internal_api_handlers.TokenPairResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGci..."
                },
                "access_token_expires_at": {
                    "type": "string",
                    "example": "2026-03-27T16:00:00Z"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGci..."
                },
                "refresh_token_expires_at": {
                    "type": "string",
                    "example": "2026-04-03T15:00:00Z"
                },
                "user": {
                    "$ref": "#/definitions/internal_api_handlers.UserView"
                }
            }
        },
...
        "internal_api_handlers.VerifyEmailConfirmBody": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440099"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for service-to-service authentication.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT access token. Example: \"Bearer eyJhbGci...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}